transfers treenode data extract functions from treemap to utilities file
